Move tagged_integer to ripple/basics
